Add filter for semantic annotations using CiTO
This filter extracts optional CiTO (Citation Typing Ontology)
information from citations and stores the information in the
document's metadata. The extracted info is intended to be used in
combination with other filters, templates, or custom writers. It
is necessary to run pandoc-citeproc *after* this filter if CiTO
data is embedded in the document; otherwise pandoc-citeproc will
interpret CiTO properties as part of the citation ID.